server: Update stats api
